moved bookings date navigation into its own component and context
also moved party bookings into its own hook, and made it subscribe, so changes (party forms being filled in) can update in realtime.

mainly did this to lay thr groundwork for moving events into its own page, with minimal work. all that has to be done is place it inside the DateNavigation on its own page, and verything will just work
