+++ firebase/src/components/Bookings/DateNavigation/DateNavigation.tsx
@@ -1,0 +1,269 @@
+import React, { FC, PropsWithChildren, ReactNode, createContext, useContext, useState } from 'react'
+import { styled } from '@mui/material/styles'
+import { useNavigate } from 'react-router-dom'
+import {
+    Grid,
+    Hidden,
+    Typography,
+    Drawer,
+    CssBaseline,
+    AppBar,
+    Button,
+    Toolbar,
+    IconButton,
+    Box,
+    LinearProgress,
+} from '@mui/material'
+import { ExitToApp as ExitToAppIcon } from '@mui/icons-material'
+import { grey } from '@mui/material/colors'
+import NavigateBefore from '@mui/icons-material/NavigateBefore'
+import NavigateNext from '@mui/icons-material/NavigateNext'
+import { DatePicker } from '@mui/x-date-pickers'
+import { StaticDatePicker } from '@mui/x-date-pickers'
+import { DateTime } from 'luxon'
+
+import * as ROUTES from '../../../constants/routes'
+import * as Logo from '../../../drawables/FizzKidzLogoHorizontal.png'
+import { useScopes } from '../../Hooks/UseScopes'
+import useFirebase from '../../Hooks/context/UseFirebase'
+
+const PREFIX = 'BookingsPage'
+
+const classes = {
+    root: `${PREFIX}-root`,
+    drawer: `${PREFIX}-drawer`,
+    drawerPaper: `${PREFIX}-drawerPaper`,
+    content: `${PREFIX}-content`,
+    appBar: `${PREFIX}-appBar`,
+    appBarToolbar: `${PREFIX}-appBarToolbar`,
+    logo: `${PREFIX}-logo`,
+    topLeft: `${PREFIX}-topLeft`,
+    topCenter: `${PREFIX}-topCenter`,
+    authTopRight: `${PREFIX}-authTopRight`,
+    noAuthTopRight: `${PREFIX}-noAuthTopRight`,
+    logoutIcon: `${PREFIX}-logoutIcon`,
+    toolbar: `${PREFIX}-toolbar`,
+    inlineDatePicker: `${PREFIX}-inlineDatePicker`,
+    location: `${PREFIX}-location`,
+    dialogueAppBar: `${PREFIX}-dialogueAppBar`,
+    divider: `${PREFIX}-divider`,
+    paper: `${PREFIX}-paper`,
+    layout: `${PREFIX}-layout`,
+    dialog: `${PREFIX}-dialog`,
+}
+
+const Root = styled('div')(({ theme }) => ({
+    [`&.${classes.root}`]: {
+        display: 'flex',
+    },
+
+    [`& .${classes.drawer}`]: {
+        width: drawerWidth,
+        flexShrink: 0,
+    },
+
+    [`& .${classes.drawerPaper}`]: {
+        width: drawerWidth,
+    },
+
+    [`& .${classes.content}`]: {
+        flexGrow: 1,
+        backgroundColor: theme.palette.background.default,
+        // padding: theme.spacing(3),
+    },
+
+    [`& .${classes.appBar}`]: {
+        zIndex: theme.zIndex.drawer + 1,
+        color: 'white',
+    },
+
+    [`& .${classes.appBarToolbar}`]: {
+        display: 'flex',
+        '@media (max-width: 550px)': {
+            justifyContent: 'space-around',
+        },
+    },
+
+    [`& .${classes.logo}`]: {
+        height: 50,
+        cursor: 'pointer',
+    },
+
+    [`& .${classes.topLeft}`]: {
+        width: '33.3%',
+        display: 'flex',
+        justifyContent: 'flex-start',
+        '@media (max-width: 550px)': {
+            display: 'none',
+        },
+    },
+
+    [`& .${classes.topCenter}`]: {
+        width: '33.3%',
+        display: 'flex',
+        justifyContent: 'center',
+    },
+
+    [`& .${classes.authTopRight}`]: {
+        width: '33.3%',
+        display: 'flex',
+        justifyContent: 'flex-end',
+        '@media (max-width: 550px)': {
+            width: 'auto',
+        },
+    },
+
+    [`& .${classes.noAuthTopRight}`]: {
+        width: '33.3%',
+        display: 'flex',
+        justifyContent: 'flex-end',
+        '@media (max-width: 550px)': {
+            display: 'none',
+        },
+    },
+
+    [`& .${classes.logoutIcon}`]: {
+        paddingTop: theme.spacing(1),
+        paddingRight: '0px',
+        paddingBottom: theme.spacing(1),
+        paddingLeft: theme.spacing(2),
+        '@media (max-width: 550px)': {
+            display: 'none',
+        },
+    },
+
+    [`& .${classes.toolbar}`]: theme.mixins.toolbar,
+}))
+
+const DateNavigationContext = createContext<{
+    date: DateTime
+    setDate: (date: DateTime) => void
+    setLoading: (loading: boolean) => void
+} | null>(null)
+
+export function useDateNavigation() {
+    const date = useContext(DateNavigationContext)
+    if (date) return date
+    throw new Error('`useDateNavigation()` must be used within a `<DateNavigation />`')
+}
+
+type WithButton = {
+    showButton: true
+    buttonLabel: string
+    onButtonPressed: () => void
+}
+
+type WithoutButton = {
+    showButton: false
+}
+
+type Props = {
+    label: string
+} & (WithButton | WithoutButton)
+
+export const DateNavigation: FC<PropsWithChildren<Props>> = (props) => {
+    const { label, showButton, children } = props
+
+    const firebase = useFirebase()
+    const navigate = useNavigate()
+    const scopes = useScopes()
+    const writePermissions = scopes.CORE === 'write'
+
+    const [date, setDate] = useState(DateTime.now())
+    const [loading, setLoading] = useState(true)
+
+    const handleDateChange = (date: DateTime) => {
+        setDate(date.set({ hour: 0, minute: 0, second: 0, millisecond: 0 }))
+    }
+
+    const handleLogout = () => {
+        firebase.doSignOut()
+    }
+
+    return (
+        <DateNavigationContext.Provider value={{ date, setDate: handleDateChange, setLoading }}>
+            <Root className={classes.root}>
+                <CssBaseline />
+                <AppBar className={classes.appBar} position="fixed">
+                    <Toolbar className={classes.appBarToolbar}>
+                        <div className={classes.topLeft}>
+                            <Typography variant="h6" color="inherit">
+                                {label}
+                            </Typography>
+                        </div>
+                        <div className={classes.topCenter}>
+                            <img
+                                className={classes.logo}
+                                src={Logo.default}
+                                onClick={() => navigate(ROUTES.LANDING)}
+                                alt="fizz kidz logo"
+                            />
+                        </div>
+                        <div className={writePermissions ? classes.authTopRight : classes.noAuthTopRight}>
+                            {writePermissions && showButton && (
+                                <Button
+                                    onClick={props.onButtonPressed}
+                                    variant="outlined"
+                                    sx={{
+                                        color: 'white',
+                                        borderColor: 'white',
+                                        '&:hover': { borderColor: 'white', background: grey[800] },
+                                    }}
+                                >
+                                    {props.buttonLabel}
+                                </Button>
+                            )}
+                            <IconButton className={classes.logoutIcon} onClick={handleLogout} size="large">
+                                <ExitToAppIcon htmlColor={'white'} />
+                            </IconButton>
+                        </div>
+                    </Toolbar>
+                </AppBar>
+                <Hidden mdDown>
+                    <Drawer
+                        className={classes.drawer}
+                        variant="permanent"
+                        classes={{
+                            paper: classes.drawerPaper,
+                        }}
+                        anchor="left"
+                    >
+                        <div className={classes.toolbar} />
+                        <StaticDatePicker
+                            value={date}
+                            slotProps={{ actionBar: { actions: ['today'] } }}
+                            onChange={(date) => handleDateChange(date!)}
+                        />
+                    </Drawer>
+                </Hidden>
+                <Grid container sx={{ marginTop: { xs: 7, sm: 8 } }}>
+                    <Box className={classes.content} sx={{ padding: { xs: 2, md: 3 } }}>
+                        <div style={{ display: 'flex', justifyContent: 'space-between' }}>
+                            <Button onClick={() => handleDateChange(date.minus({ days: 1 }))}>
+                                <NavigateBefore />
+                            </Button>
+                            <DatePicker
+                                closeOnSelect
+                                value={date}
+                                slotProps={{
+                                    textField: { sx: { input: { textAlign: 'center' } }, fullWidth: true },
+                                    actionBar: { actions: ['today'] },
+                                }}
+                                format="ccc, LLL d, y"
+                                onChange={(date) => date && handleDateChange(date)}
+                            />
+                            <Button onClick={() => handleDateChange(date.plus({ days: 1 }))}>
+                                <NavigateNext />
+                            </Button>
+                        </div>
+                        {loading && <LinearProgress style={{ marginTop: 9 }} color="secondary" />}
+                        {!loading && <div style={{ height: 13 }} />}
+                        {children}
+                    </Box>
+                </Grid>
+            </Root>
+        </DateNavigationContext.Provider>
+    )
+}
+
+const drawerWidth = 320
