+++ firebase/src/components/Bookings/Events/UseEvents.tsx
@@ -2,11 +2,12 @@ import { useEffect, useState } from 'react'
 import useFirebase from '../../Hooks/context/UseFirebase'
 import { Service, EventBooking } from 'fizz-kidz'
 import { convertTimestamps } from '../../../utilities/firebase/converters'
+import { useDateNavigation } from '../DateNavigation/DateNavigation'
 
-export function useEvents(_date: Date = new Date()) {
+export function useEvents() {
     const firebase = useFirebase()
 
-    const [date, setDate] = useState(_date)
+    const { date } = useDateNavigation()
     const [events, setEvents] = useState<Service<EventBooking[]>>({ status: 'loading' })
 
     useEffect(() => {
@@ -14,19 +15,15 @@ export function useEvents(_date: Date = new Date()) {
             // since we need to get a range between startDate and endDate,
             // and firestore does not support equality operators '>', '<' on multiple fields,
             // fetch all events that may have already started, and filter the rest on the frontend
-
-            date.setHours(0, 0, 0, 0)
-            const nextDay = new Date(date.getTime())
-            nextDay.setDate(nextDay.getDate() + 1)
+            const nextDay = date.plus({ days: 1 })
             // an event will never be 90 days long, so a safe window
-            const ninetyDaysAgo = new Date(date.getTime())
-            ninetyDaysAgo.setDate(ninetyDaysAgo.getDate() - 90)
+            const ninetyDaysAgo = date.minus({ days: 90 })
 
             try {
                 const snap = await firebase.db
                     .collection('events')
-                    .where('startTime', '<', nextDay)
-                    .where('startTime', '>', ninetyDaysAgo)
+                    .where('startTime', '<', nextDay.toJSDate())
+                    .where('startTime', '>', ninetyDaysAgo.toJSDate())
                     .get()
 
                 const events = snap.docs
@@ -34,7 +31,7 @@ export function useEvents(_date: Date = new Date()) {
                         const data = doc.data()
                         if (data) {
                             const event = convertTimestamps(data) as EventBooking
-                            if (event.endTime > date) {
+                            if (event.endTime > date.toJSDate()) {
                                 return event
                             }
                         }
@@ -52,5 +49,5 @@ export function useEvents(_date: Date = new Date()) {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [date])
 
-    return [events, setDate] as const
+    return events
 }
